Old sync manager removed and some tests updated
