+++ src/main/java/io/split/android/client/service/synchronizer/SyncManager.java
@@ -5,6 +5,7 @@ import androidx.annotation.VisibleForTesting;
 import io.split.android.client.dtos.Event;
 import io.split.android.client.impressions.Impression;
 
+// TODO: Will be renamed to SyncManager on final integration
 @VisibleForTesting(otherwise = VisibleForTesting.PROTECTED)
 public interface SyncManager {
     void start();
@@ -13,11 +14,12 @@ public interface SyncManager {
 
     void resume();
 
-    void stop();
-
     void flush();
 
     void pushEvent(Event event);
 
     void pushImpression(Impression impression);
+
+    void stop();
+
 }
